@@ -8,7 +8,7 @@
     "lint": "yarn lint:code && yarn lint:docs",
     "lint:code": "eslint --ignore-path .gitignore --ext .js,.ts,.tsx .",
     "lint:docs": "alex app/pages/docs/**/*.mdx",
-    "english-slugify": "node scripts/english-slugify.js",
+    "english-slugify": "node scripts/english-slugify.mjs",
     "prepare": "husky install"
   },
   "lint-staged": {
@@ -18,18 +18,18 @@
   },
   "dependencies": {
     "@docsearch/react": "1.0.0-alpha.28",
-    "@octokit/rest": "18.5.2",
-    "@reach/rect": "0.14.0",
-    "@sindresorhus/slugify": "1.1.0",
+    "@octokit/rest": "18.5.3",
+    "@reach/rect": "0.15.0",
+    "@sindresorhus/slugify": "2.0.0",
     "@visx/hierarchy": "1.7.0",
     "@visx/responsive": "1.7.0",
-    "blitz": "0.33.1",
+    "blitz": "0.34.3",
     "clsx": "1.1.1",
     "dlv": "1.1.3",
     "fathom-client": "3.0.0",
     "focus-visible": "5.2.0",
-    "framer-motion": "4.0.3",
-    "gray-matter": "4.0.2",
+    "framer-motion": "4.1.9",
+    "gray-matter": "4.0.3",
     "just-group-by": "1.0.0",
     "next-themes": "0.0.14",
     "prismjs": "1.23.0",
@@ -48,27 +48,26 @@
   },
   "devDependencies": {
     "@mdx-js/loader": "1.6.22",
-    "@next/bundle-analyzer": "10.0.9",
+    "@next/bundle-analyzer": "10.1.3",
     "@svgr/webpack": "5.5.0",
-    "@tailwindcss/jit": "0.1.17",
     "@tailwindcss/typography": "0.4.0",
     "alex": "9.1.0",
     "autoprefixer": "10.2.5",
     "babel-plugin-preval": "5.0.0",
-    "eslint": "7.23.0",
+    "eslint": "7.25.0",
     "eslint-plugin-simple-import-sort": "7.0.0",
     "file-loader": "6.2.0",
     "glob": "7.1.6",
     "husky": "6.0.0",
     "lint-staged": "10.5.4",
     "minimatch": "3.0.4",
-    "postcss": "8.2.8",
+    "postcss": "8.2.12",
     "postcss-nested": "5.0.5",
     "prettier": "2.2.1",
     "pretty-quick": "3.1.0",
     "remark-admonitions": "1.2.1",
     "simple-functional-loader": "1.2.1",
-    "tailwindcss": "2.0.4"
+    "tailwindcss": "2.1.2"
   },
   "private": true
 }
