@@ -1,4 +1,13 @@
 // Use a "maybe" level of profanity instead of the default "unlikely".
 exports.profanitySureness = 1
 
-exports.allow = ["simple", "special", "invalid", "he-she", "her-him", "herself-himself", "obvious"]
+exports.allow = [
+  "simple",
+  "special",
+  "invalid",
+  "he-she",
+  "her-him",
+  "herself-himself",
+  "obvious",
+  "easy",
+]
