@@ -7,11 +7,11 @@ import tokenize from "../../macros/tokenize.macro.js"
 
 const pageTokenized = tokenize.jsx(
   `// app/pages/projects/new.tsx
-import { Link, useRouter, useMutation, BlitzPage } from "blitz"
+import { Link, Routes, useRouter, useMutation, BlitzPage } from "blitz"
 import Layout from "app/core/layouts/Layout"
 // Notice how we import the server function directly
 import createProject, {CreateProject} from "app/projects/mutations/createProject"
-import { ProjectForm, FORM_ERROR } from "app/projects/components/ProjectForm"
+import { ProjectForm } from "app/projects/components/ProjectForm"
 
 const NewProjectPage: BlitzPage = () => {
   const router = useRouter()
@@ -25,13 +25,10 @@ const NewProjectPage: BlitzPage = () => {
         submitText="Create Project"
         schema={CreateProject}
         onSubmit={async (values) => {
-          try {
-            // This is equivalent to calling the server function directly
-            const project = await createProjectMutation(values)
-            router.push("/projects/" + project.id)
-          } catch (error) {
-            return { [FORM_ERROR]: error.toString() }
-          }
+          // This is equivalent to calling the server function directly
+          const project = await createProjectMutation(values)
+          // Notice the 'Routes' object Blitz provides for routing
+          router.push(Routes.ProjectsPage({projectId: project.id}}))
         }}
       />
     </div>
