@@ -1,3 +1,3 @@
 module.exports = {
-  plugins: ["@tailwindcss/jit", "postcss-nested", "autoprefixer"],
+  plugins: ["tailwindcss", "postcss-nested", "autoprefixer"],
 }
